+++ CHANGELOG.md
@@ -8,6 +8,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Changed
 
 - Using absolute instead of relative imports ([[#995](https://github.com/pdfminer/pdfminer.six/pull/995)])
+- Allow suppression of `(cid:N)` in `pdf2txt.py` ([#1070](https://github.com/pdfminer/pdfminer.six/pull/1070))
 
 ### Deprecated